+++ src/advanced/index.ts
@@ -9,7 +9,7 @@ export { GlyphMesh } from "../render/GlyphMesh";
 export { RenderCoordinator } from "../render/RenderCoordinator";
 export { WebGLAdapter } from "../render/WebGLAdapter";
 export { WebGPUAdapter } from "../render/WebGPUAdapter";
-export { GLYPH_INSTANCE_STRIDE } from "../render/types";
+export { GLYPH_INSTANCE_STRIDE, GLYPH_TEXTURE_BANK_SIZE } from "../render/types";
 export { TRANSFORM_PALETTE_STRIDE, TransformPalette } from "../render/TransformPalette";
 export type {
   AtlasCommit,